+++ Common/Services/Integrations/FauxCore/IFauxCoreApi.cs
@@ -3,17 +3,18 @@ namespace StardewMods.Common.Services.Integrations.FauxCore;
 /// <summary>Api for shared functionality between mods.</summary>
 public interface IFauxCoreApi
 {
-    /// <summary>Create an instance of the ILog service for the given IMonitor.</summary>
-    /// <param name="monitor">Dependency used for monitoring and logging.</param>
-    /// <returns>An instance of ILog that is associated with the provided IMonitor.</returns>
-    public ILog CreateLogService(IMonitor monitor);
+    /// <summary>Gets an instance of the expression handler service.</summary>
+    public IExpressionHandler ExpressionHandler { get; }
 
-    /// <summary>Create an instance of the IPatchManager service.</summary>
-    /// <returns>An instance of IPatchManager.</returns>
-    /// <param name="log">Dependency used for logging debug information to the console.</param>
-    public IPatchManager CreatePatchService(ILog log);
+    /// <summary>Gets or sets the <see cref="IMonitor" /> instance.</summary>
+    public IMonitor? Monitor { get; set; }
 
-    /// <summary>Create an instance of the IThemeHelper service.</summary>
-    /// <returns>An instance of IThemeHelper.</returns>
-    public IThemeHelper CreateThemeService();
+    /// <summary>Gets an instance of the log service.</summary>
+    public ILog Log { get; }
+
+    /// <summary>Gets an instance of the patch manager service.</summary>
+    public IPatchManager PatchManager { get; }
+
+    /// <summary>Gets an instance of the theme helper service.</summary>
+    public IThemeHelper ThemeHelper { get; }
 }