+++ ExpandedStorage/ModEntry.cs
@@ -41,9 +41,9 @@ public sealed class ModEntry : Mod
         this.container.RegisterSingleton<ContentPatcherIntegration>();
         this.container.RegisterSingleton<IEventManager, EventManager>();
         this.container.RegisterSingleton<FauxCoreIntegration>();
-        this.container.RegisterSingleton<ILog, Logger>();
+        this.container.RegisterSingleton<ILog, FauxCoreIntegration>();
         this.container.RegisterSingleton<ModPatches>();
-        this.container.RegisterSingleton<IPatchManager, Patcher>();
+        this.container.RegisterSingleton<IPatchManager, FauxCoreIntegration>();
         this.container.RegisterSingleton<GenericModConfigMenuIntegration>();
         this.container.RegisterSingleton<ToolbarIconsIntegration>();