+++ FauxCore/Framework/FauxCoreApi.cs
@@ -9,25 +9,44 @@ public sealed class FauxCoreApi : IFauxCoreApi
 {
     private readonly Func<IModConfig> getConfig;
     private readonly IModInfo modInfo;
-    private readonly IThemeHelper themeHelper;
+
+    private ILog? log;
+    private IPatchManager? patchManager;
 
     /// <summary>Initializes a new instance of the <see cref="FauxCoreApi" /> class.</summary>
     /// <param name="modInfo">Dependency used for accessing mod info.</param>
+    /// <param name="expressionHandler">Dependency used for parsing expressions.</param>
     /// <param name="getConfig">Dependency used for accessing config data.</param>
     /// <param name="themeHelper">Dependency used for swapping palettes.</param>
-    public FauxCoreApi(IModInfo modInfo, Func<IModConfig> getConfig, IThemeHelper themeHelper)
+    public FauxCoreApi(
+        IModInfo modInfo,
+        IExpressionHandler expressionHandler,
+        Func<IModConfig> getConfig,
+        IThemeHelper themeHelper)
     {
         this.modInfo = modInfo;
+        this.ExpressionHandler = expressionHandler;
         this.getConfig = getConfig;
-        this.themeHelper = themeHelper;
+        this.ThemeHelper = themeHelper;
     }
 
-    /// <inheritdoc />
-    public ILog CreateLogService(IMonitor monitor) => new Log(this.getConfig, monitor);
+    public IExpressionHandler ExpressionHandler { get; }
 
     /// <inheritdoc />
-    public IPatchManager CreatePatchService(ILog log) => new PatchManager(log, this.modInfo.Manifest);
+    public ILog Log =>
+        this.log ??= new Log(
+            this.getConfig,
+            this.Monitor ?? throw new InvalidOperationException("Monitor is not set."));
 
     /// <inheritdoc />
-    public IThemeHelper CreateThemeService() => this.themeHelper;
+    public IPatchManager PatchManager =>
+        this.patchManager ??= new PatchManager(
+            this.Log ?? throw new InvalidOperationException("Log is not set."),
+            this.modInfo.Manifest);
+
+    /// <inheritdoc />
+    public IThemeHelper ThemeHelper { get; }
+
+    /// <inheritdoc />
+    public IMonitor? Monitor { get; set; }
 }