+++ FauxCore/ModEntry.cs
@@ -37,6 +37,7 @@ public sealed class ModEntry : Mod
         this.container.RegisterSingleton<IModConfig, ConfigManager>();
         this.container.RegisterSingleton<ConfigManager, ConfigManager>();
         this.container.RegisterSingleton<IEventManager, EventManager>();
+        this.container.RegisterSingleton<IExpressionHandler, ExpressionHandler>();
         this.container.RegisterSingleton<GenericModConfigMenuIntegration>();
         this.container.RegisterSingleton<ILog, Log>();
         this.container.RegisterSingleton<IThemeHelper, ThemeHelper>();
@@ -49,6 +50,7 @@ public sealed class ModEntry : Mod
     public override object GetApi(IModInfo mod) =>
         new FauxCoreApi(
             mod,
+            this.container.GetInstance<IExpressionHandler>(),
             this.container.GetInstance<Func<IModConfig>>(),
             this.container.GetInstance<IThemeHelper>());