+++ ToolbarIcons/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changed
 
+* Updated for FauxCore 1.2.0.
 * If config file is missing, it will attempt to restore from global data.
 
 ## 2.7.2 (April 12, 2024)